pyproject.toml
[project]
name = "crypto-eye"
version = "0.1.0"
description = "360 Crypto Eye - modular crypto intelligence stack"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-telegram-bot>=21.4",
  "httpx>=0.27.0",
  "pandas>=2.2.2",
  "numpy>=2.0.1",
  "plotly>=5.23.0",
  "kaleido>=0.2.1",
  "sqlalchemy>=2.0.32",
  "psycopg[binary]>=3.2.1",
  "redis>=5.0.8",
  "apscheduler>=3.10.4",
  "tenacity>=9.0.0",
  "python-dotenv>=1.0.1",
]
